--- a/backend/groth16/groth16.go
+++ b/backend/groth16/groth16.go
@@ -29,6 +29,7 @@ import (
 	backend_bls12377 "github.com/consensys/gnark/internal/backend/bls12-377/cs"
 	backend_bls12381 "github.com/consensys/gnark/internal/backend/bls12-381/cs"
 	backend_bls24315 "github.com/consensys/gnark/internal/backend/bls24-315/cs"
+	backend_bls24317 "github.com/consensys/gnark/internal/backend/bls24-317/cs"
 	backend_bn254 "github.com/consensys/gnark/internal/backend/bn254/cs"
 	backend_bw6633 "github.com/consensys/gnark/internal/backend/bw6-633/cs"
 	backend_bw6761 "github.com/consensys/gnark/internal/backend/bw6-761/cs"
@@ -36,6 +37,7 @@ import (
 	witness_bls12377 "github.com/consensys/gnark/internal/backend/bls12-377/witness"
 	witness_bls12381 "github.com/consensys/gnark/internal/backend/bls12-381/witness"
 	witness_bls24315 "github.com/consensys/gnark/internal/backend/bls24-315/witness"
+	witness_bls24317 "github.com/consensys/gnark/internal/backend/bls24-317/witness"
 	witness_bn254 "github.com/consensys/gnark/internal/backend/bn254/witness"
 	witness_bw6633 "github.com/consensys/gnark/internal/backend/bw6-633/witness"
 	witness_bw6761 "github.com/consensys/gnark/internal/backend/bw6-761/witness"
@@ -45,6 +47,7 @@ import (
 	groth16_bls12377 "github.com/consensys/gnark/internal/backend/bls12-377/groth16"
 	groth16_bls12381 "github.com/consensys/gnark/internal/backend/bls12-381/groth16"
 	groth16_bls24315 "github.com/consensys/gnark/internal/backend/bls24-315/groth16"
+	groth16_bls24317 "github.com/consensys/gnark/internal/backend/bls24-317/groth16"
 	groth16_bn254 "github.com/consensys/gnark/internal/backend/bn254/groth16"
 	groth16_bw6633 "github.com/consensys/gnark/internal/backend/bw6-633/groth16"
 	groth16_bw6761 "github.com/consensys/gnark/internal/backend/bw6-761/groth16"
@@ -133,6 +136,12 @@ func Verify(proof Proof, vk VerifyingKey, publicWitness *witness.Witness) error 
 			return witness.ErrInvalidWitness
 		}
 		return groth16_bw6761.Verify(_proof, vk.(*groth16_bw6761.VerifyingKey), *w)
+	case *groth16_bls24317.Proof:
+		w, ok := publicWitness.Vector.(*witness_bls24317.Witness)
+		if !ok {
+			return witness.ErrInvalidWitness
+		}
+		return groth16_bls24317.Verify(_proof, vk.(*groth16_bls24317.VerifyingKey), *w)
 	case *groth16_bls24315.Proof:
 		w, ok := publicWitness.Vector.(*witness_bls24315.Witness)
 		if !ok {
@@ -189,6 +198,12 @@ func Prove(r1cs frontend.CompiledConstraintSystem, pk ProvingKey, fullWitness *w
 			return nil, witness.ErrInvalidWitness
 		}
 		return groth16_bw6761.Prove(_r1cs, pk.(*groth16_bw6761.ProvingKey), *w, opt)
+	case *backend_bls24317.R1CS:
+		w, ok := fullWitness.Vector.(*witness_bls24317.Witness)
+		if !ok {
+			return nil, witness.ErrInvalidWitness
+		}
+		return groth16_bls24317.Prove(_r1cs, pk.(*groth16_bls24317.ProvingKey), *w, opt)
 	case *backend_bls24315.R1CS:
 		w, ok := fullWitness.Vector.(*witness_bls24315.Witness)
 		if !ok {
@@ -245,6 +260,13 @@ func Setup(r1cs frontend.CompiledConstraintSystem) (ProvingKey, VerifyingKey, er
 			return nil, nil, err
 		}
 		return &pk, &vk, nil
+	case *backend_bls24317.R1CS:
+		var pk groth16_bls24317.ProvingKey
+		var vk groth16_bls24317.VerifyingKey
+		if err := groth16_bls24317.Setup(_r1cs, &pk, &vk); err != nil {
+			return nil, nil, err
+		}
+		return &pk, &vk, nil
 	case *backend_bls24315.R1CS:
 		var pk groth16_bls24315.ProvingKey
 		var vk groth16_bls24315.VerifyingKey
@@ -292,6 +314,12 @@ func DummySetup(r1cs frontend.CompiledConstraintSystem) (ProvingKey, error) {
 			return nil, err
 		}
 		return &pk, nil
+	case *backend_bls24317.R1CS:
+		var pk groth16_bls24317.ProvingKey
+		if err := groth16_bls24317.DummySetup(_r1cs, &pk); err != nil {
+			return nil, err
+		}
+		return &pk, nil
 	case *backend_bls24315.R1CS:
 		var pk groth16_bls24315.ProvingKey
 		if err := groth16_bls24315.DummySetup(_r1cs, &pk); err != nil {
@@ -322,6 +350,8 @@ func NewProvingKey(curveID ecc.ID) ProvingKey {
 		pk = &groth16_bls12381.ProvingKey{}
 	case ecc.BW6_761:
 		pk = &groth16_bw6761.ProvingKey{}
+	case ecc.BLS24_317:
+		pk = &groth16_bls24317.ProvingKey{}
 	case ecc.BLS24_315:
 		pk = &groth16_bls24315.ProvingKey{}
 	case ecc.BW6_633:
@@ -345,6 +375,8 @@ func NewVerifyingKey(curveID ecc.ID) VerifyingKey {
 		vk = &groth16_bls12381.VerifyingKey{}
 	case ecc.BW6_761:
 		vk = &groth16_bw6761.VerifyingKey{}
+	case ecc.BLS24_317:
+		vk = &groth16_bls24317.VerifyingKey{}
 	case ecc.BLS24_315:
 		vk = &groth16_bls24315.VerifyingKey{}
 	case ecc.BW6_633:
@@ -369,6 +401,8 @@ func NewProof(curveID ecc.ID) Proof {
 		proof = &groth16_bls12381.Proof{}
 	case ecc.BW6_761:
 		proof = &groth16_bw6761.Proof{}
+	case ecc.BLS24_317:
+		proof = &groth16_bls24317.Proof{}
 	case ecc.BLS24_315:
 		proof = &groth16_bls24315.Proof{}
 	case ecc.BW6_633:
@@ -393,6 +427,8 @@ func NewCS(curveID ecc.ID) frontend.CompiledConstraintSystem {
 		r1cs = &backend_bls12381.R1CS{}
 	case ecc.BW6_761:
 		r1cs = &backend_bw6761.R1CS{}
+	case ecc.BLS24_317:
+		r1cs = &backend_bls24317.R1CS{}
 	case ecc.BLS24_315:
 		r1cs = &backend_bls24315.R1CS{}
 	case ecc.BW6_633: