--- a/frontend/cs/r1cs/builder.go
+++ b/frontend/cs/r1cs/builder.go
@@ -37,6 +37,7 @@ import (
 	bls12377r1cs "github.com/consensys/gnark/internal/backend/bls12-377/cs"
 	bls12381r1cs "github.com/consensys/gnark/internal/backend/bls12-381/cs"
 	bls24315r1cs "github.com/consensys/gnark/internal/backend/bls24-315/cs"
+	bls24317r1cs "github.com/consensys/gnark/internal/backend/bls24-317/cs"
 	bn254r1cs "github.com/consensys/gnark/internal/backend/bn254/cs"
 	bw6633r1cs "github.com/consensys/gnark/internal/backend/bw6-633/cs"
 	bw6761r1cs "github.com/consensys/gnark/internal/backend/bw6-761/cs"
@@ -390,6 +391,8 @@ func (cs *r1cs) Compile() (frontend.CompiledConstraintSystem, error) {
 		return bw6633r1cs.NewR1CS(res, cs.st.Coeffs), nil
 	case ecc.BLS24_315:
 		return bls24315r1cs.NewR1CS(res, cs.st.Coeffs), nil
+	case ecc.BLS24_317:
+		return bls24317r1cs.NewR1CS(res, cs.st.Coeffs), nil
 	default:
 		panic("not implemtented")
 	}