--- a/std/algebra/twistededwards/twistededwards.go
+++ b/std/algebra/twistededwards/twistededwards.go
@@ -25,6 +25,7 @@ import (
 	edbls12381_bandersnatch "github.com/consensys/gnark-crypto/ecc/bls12-381/bandersnatch"
 	edbls12381 "github.com/consensys/gnark-crypto/ecc/bls12-381/twistededwards"
 	edbls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315/twistededwards"
+	edbls24317 "github.com/consensys/gnark-crypto/ecc/bls24-317/twistededwards"
 	edbn254 "github.com/consensys/gnark-crypto/ecc/bn254/twistededwards"
 	edbw6633 "github.com/consensys/gnark-crypto/ecc/bw6-633/twistededwards"
 	edbw6761 "github.com/consensys/gnark-crypto/ecc/bw6-761/twistededwards"
@@ -104,6 +105,8 @@ func GetCurveParams(id twistededwards.ID) (*CurveParams, error) {
 		params = newEdBLS12_381()
 	case twistededwards.BLS12_381_BANDERSNATCH:
 		params = newEdBLS12_381_BANDERSNATCH()
+	case twistededwards.BLS24_317:
+		params = newEdBLS24_317()
 	case twistededwards.BLS24_315:
 		params = newEdBLS24_315()
 	case twistededwards.BW6_761:
@@ -125,6 +128,8 @@ func GetSnarkCurve(id twistededwards.ID) (ecc.ID, error) {
 		return ecc.BLS12_377, nil
 	case twistededwards.BLS12_381, twistededwards.BLS12_381_BANDERSNATCH:
 		return ecc.BLS12_381, nil
+	case twistededwards.BLS24_317:
+		return ecc.BLS24_317, nil
 	case twistededwards.BLS24_315:
 		return ecc.BLS24_315, nil
 	case twistededwards.BW6_761:
@@ -226,6 +231,21 @@ func newEdBW6_633() *CurveParams {
 func newEdBW6_761() *CurveParams {
 
 	edcurve := edbw6761.GetEdwardsCurve()
+
+	r := newCurveParams()
+	edcurve.A.ToBigIntRegular(r.A)
+	edcurve.D.ToBigIntRegular(r.D)
+	edcurve.Cofactor.ToBigIntRegular(r.Cofactor)
+	r.Order.Set(&edcurve.Order)
+	edcurve.Base.X.ToBigIntRegular(r.Base[0])
+	edcurve.Base.Y.ToBigIntRegular(r.Base[1])
+	return r
+
+}
+
+func newEdBLS24_317() *CurveParams {
+
+	edcurve := edbls24317.GetEdwardsCurve()
 
 	r := newCurveParams()
 	edcurve.A.ToBigIntRegular(r.A)