--- a/std/hash/mimc/encrypt.go
+++ b/std/hash/mimc/encrypt.go
@@ -21,6 +21,7 @@ import (
 	bls12377 "github.com/consensys/gnark-crypto/ecc/bls12-377/fr/mimc"
 	bls12381 "github.com/consensys/gnark-crypto/ecc/bls12-381/fr/mimc"
 	bls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315/fr/mimc"
+	bls24317 "github.com/consensys/gnark-crypto/ecc/bls24-317/fr/mimc"
 	bn254 "github.com/consensys/gnark-crypto/ecc/bn254/fr/mimc"
 	bw6633 "github.com/consensys/gnark-crypto/ecc/bw6-633/fr/mimc"
 	bw6761 "github.com/consensys/gnark-crypto/ecc/bw6-761/fr/mimc"
@@ -39,6 +40,7 @@ func init() {
 	encryptFuncs[ecc.BW6_761] = encryptPow5
 	encryptFuncs[ecc.BW6_633] = encryptPow5
 	encryptFuncs[ecc.BLS24_315] = encryptPow5
+	encryptFuncs[ecc.BLS24_317] = encryptPow7
 
 	newMimc = make(map[ecc.ID]func(frontend.API) MiMC)
 	newMimc[ecc.BN254] = newMimcBN254
@@ -47,6 +49,7 @@ func init() {
 	newMimc[ecc.BW6_761] = newMimcBW761
 	newMimc[ecc.BW6_633] = newMimcBW633
 	newMimc[ecc.BLS24_315] = newMimcBLS315
+	newMimc[ecc.BLS24_317] = newMimcBLS317
 }
 
 // -------------------------------------------------------------------------------------------------
@@ -88,6 +91,15 @@ func newMimcBW761(api frontend.API) MiMC {
 	return res
 }
 
+func newMimcBLS317(api frontend.API) MiMC {
+	res := MiMC{}
+	res.params = bls24317.GetConstants()
+	res.id = ecc.BLS24_317
+	res.h = 0
+	res.api = api
+	return res
+}
+
 func newMimcBLS315(api frontend.API) MiMC {
 	res := MiMC{}
 	res.params = bls24315.GetConstants()
@@ -115,12 +127,29 @@ func pow5(api frontend.API, x frontend.Variable) frontend.Variable {
 	return api.Mul(r, x)
 }
 
+func pow7(api frontend.API, x frontend.Variable) frontend.Variable {
+	r := api.Mul(x, x)
+	r = api.Mul(r, r)
+	r = api.Mul(r, r)
+	return api.DivUnchecked(r, x)
+}
+
 // encryptBn256 of a mimc run expressed as r1cs
 // m is the message, k the key
 func encryptPow5(h MiMC, m frontend.Variable) frontend.Variable {
 	x := m
 	for i := 0; i < len(h.params); i++ {
 		x = pow5(h.api, h.api.Add(x, h.h, h.params[i]))
+	}
+	return h.api.Add(x, h.h)
+}
+
+// encryptBLS24317 of a mimc run expressed as r1cs
+// m is the message, k the key
+func encryptPow7(h MiMC, m frontend.Variable) frontend.Variable {
+	x := m
+	for i := 0; i < len(h.params); i++ {
+		x = pow7(h.api, h.api.Add(x, h.h, h.params[i]))
 	}
 	return h.api.Add(x, h.h)
 }