… README.md
… backend/witness/witness_test.go
… internal/backend/bls24-317/cs/r1cs_test.go
… internal/backend/bls24-317/groth16/groth16_test.go
… internal/backend/bls24-317/groth16/marshal.go
… internal/backend/bls24-317/groth16/marshal_test.go
… internal/backend/bls24-317/groth16/prove.go
… internal/backend/bls24-317/groth16/setup.go
… internal/backend/bls24-317/groth16/verify.go
… internal/backend/bls24-317/plonk/marshal.go
… internal/backend/bls24-317/plonk/marshal_test.go
… internal/backend/bls24-317/plonk/plonk_test.go
… internal/backend/bls24-317/plonk/prove.go
… internal/backend/bls24-317/plonk/setup.go
… internal/backend/bls24-317/plonk/verify.go
… internal/stats/latest.stats
… std/algebra/twistededwards/curve_test.go
… std/hash/mimc/mimc_test.go
… std/signature/eddsa/eddsa_test.go
… test/kzg_srs.go