Optimized curve for KZG-Plonk (faster G1 MSM, `2^60|r-1`). Compared to BLS12-381, for a Plonk circuit of 40k constraints:

```
benchmark                  old ns/op       new ns/op       delta
BenchmarkSetup/setup-8     10815314958     7982441375      -26.19%
BenchmarkProver-8          14243060417     13210455125     -7.25%
BenchmarkVerifier-8        13538656        14977696        +10.63%
```