@yelhousni 2022-05-06
needs gnark-crypto PR to be merged first (https://github.com/ConsenSys/gnark-crypto/pull/179)